컴파일 에러 수정, CameraController추가, GlobalBuffer에 WVP 행렬 추가

+++ Libraries/Include/Engine/GameObject.h
@@ -43,7 +43,7 @@ public:
 		eComponentType type = C::GetType();
 		if (type != eComponentType::Script)
 		{
-			_fixedComponent[type] = component;
+			_fixedComponent[(int)type] = component;
 		}
 		else
 		{
